fix call adapter issue
