+++ app/src/main/java/com/technopark/youtrader/di/AppModule.kt
@@ -9,7 +9,7 @@ import com.technopark.youtrader.database.AppDatabase
 import com.technopark.youtrader.network.CryptoCurrencyApi
 import com.technopark.youtrader.network.FirebaseService
 import com.technopark.youtrader.network.IAuthService
-import com.technopark.youtrader.network.NetworkResponseAdapterFactory
+import com.technopark.youtrader.network.retrofit.NetworkResponseAdapterFactory
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -17,7 +17,6 @@ import dagger.hilt.components.SingletonComponent
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
-//import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.converter.moshi.MoshiConverterFactory
 import javax.inject.Singleton
