+++ app/src/main/java/com/technopark/youtrader/network/CryptoCurrencyNetworkService.kt
@@ -1,11 +1,7 @@
 package com.technopark.youtrader.network
 
 import android.util.Log
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import com.technopark.youtrader.model.CryptoCurrencyExample
-import com.technopark.youtrader.model.CryptoCurrencyWrapper
-import com.technopark.youtrader.repository.CryptoCurrencyRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
@@ -29,21 +25,12 @@ class CryptoCurrencyNetworkService @Inject constructor(private val cryptoApi: Cr
 
     fun getCurrency() {
         GlobalScope.launch {
-//            withContext(Dispatchers.IO) {
-//                try {
-//                    val response = cryptoApi.getCryptoCurrency("bitcoin")
-////                    Log.d(TAG, response.data.getOrDefault("name", "some string"))
-//                    Log.d(TAG, response.data.name)
-//                } catch (e: Exception) {
-//                    Log.d(TAG, "Error")
-//                }
-//            }
-
             withContext(Dispatchers.IO) {
                 val response = cryptoApi.getCurrency("bitcoin")
                 when (response) {
-                    is NetworkResponse.Success -> Log.d(TAG, response.body.name)
-//                    is NetworkResponse.ApiError -> Log.d(TAG, "ApiError ${response.body.message}")
+                    is NetworkResponse.Success -> Log.d(TAG, response.body.data.name)
+                    is NetworkResponse.ApiError ->
+                        Log.d(TAG, "ApiError. message: ${response.error}, code: ${response.code}")
                     is NetworkResponse.NetworkError -> Log.d(TAG, "NetworkError")
                     is NetworkResponse.UnknownError -> Log.d(TAG, "UnknownError")
                 }
