+++ app/src/main/java/com/technopark/youtrader/network/retrofit/NetworkResponseAdapter.kt
@@ -1,5 +1,6 @@
-package com.technopark.youtrader.network
+package com.technopark.youtrader.network.retrofit
 
+import com.technopark.youtrader.network.NetworkResponse
 import retrofit2.Call
 import retrofit2.CallAdapter
 import java.lang.reflect.Type
