+++ app/src/main/java/com/technopark/youtrader/network/retrofit/NetworkResponseCall.kt
@@ -1,0 +1,88 @@
+package com.technopark.youtrader.network.retrofit
+
+import com.technopark.youtrader.network.NetworkResponse
+import okhttp3.Request
+import okio.Timeout
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import java.io.IOException
+
+internal class NetworkResponseCall<S : Any>(
+    private val delegate: Call<S>,
+) : Call<NetworkResponse<S>> {
+    override fun clone(): Call<NetworkResponse<S>> {
+        return NetworkResponseCall(delegate.clone())
+    }
+
+    override fun execute(): Response<NetworkResponse<S>> {
+        throw UnsupportedOperationException("NetworkResponseCall doesn't support execute")
+    }
+
+    override fun enqueue(callback: Callback<NetworkResponse<S>>) {
+        return delegate.enqueue(
+            object : Callback<S> {
+                override fun onResponse(call: Call<S>, response: Response<S>) {
+                    val body = response.body()
+                    val code = response.code()
+                    val error = response.errorBody()
+
+                    if (response.isSuccessful) {
+                        if (body != null) {
+                            callback.onResponse(
+                                this@NetworkResponseCall,
+                                Response.success(NetworkResponse.Success(body))
+                            )
+                        } else {
+                            // Response is successful but the body is null
+                            callback.onResponse(
+                                this@NetworkResponseCall,
+                                Response.success(NetworkResponse.UnknownError(null))
+                            )
+                        }
+                    } else {
+                        if (error != null) {
+                            callback.onResponse(
+                                this@NetworkResponseCall,
+                                Response.success(NetworkResponse.ApiError(error.string(), code))
+                            )
+                        } else {
+                            callback.onResponse(
+                                this@NetworkResponseCall,
+                                Response.success(NetworkResponse.UnknownError(null))
+                            )
+                        }
+                    }
+                }
+
+                override fun onFailure(call: Call<S>, throwable: Throwable) {
+                    val networkResponse = when (throwable) {
+                        is IOException -> NetworkResponse.NetworkError(throwable)
+                        else -> NetworkResponse.UnknownError(throwable)
+                    }
+                    callback.onResponse(this@NetworkResponseCall, Response.success(networkResponse))
+                }
+            }
+        )
+    }
+
+    override fun isExecuted(): Boolean {
+        return delegate.isExecuted
+    }
+
+    override fun cancel() {
+        delegate.cancel()
+    }
+
+    override fun isCanceled(): Boolean {
+        return delegate.isCanceled
+    }
+
+    override fun request(): Request {
+        return delegate.request()
+    }
+
+    override fun timeout(): Timeout {
+        return delegate.timeout()
+    }
+}
