+++ app/src/main/java/com/technopark/youtrader/ui/auth/AuthViewModel.kt
@@ -2,14 +2,11 @@ package com.technopark.youtrader.ui.auth
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.viewModelScope
 import com.technopark.youtrader.base.BaseViewModel
 import com.technopark.youtrader.model.CryptoCurrencyExample
 import com.technopark.youtrader.network.IAuthService
 import com.technopark.youtrader.repository.CryptoCurrencyRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
@@ -23,15 +20,6 @@ class AuthViewModel @Inject constructor(
 
     init {
         repository.getCurrency()
-    }
-
-    fun getCryptoCurrencies() {
-//        viewModelScope.launch {
-//            repository.getCurrencies()
-//                .collect { cryptoCurrencies ->
-//                    _cryptoCurrencies.value = cryptoCurrencies
-//                }
-//        }
     }
 
     fun signUp(email: String, password: String) = authService.sighUp(email, password)
